:recycle: Moving GroupByPURL to the internal/utility/purl package and rename it to Group to match conventions
